ProVision/requirements.txt
torch
transformers
Pillow
Gemini AI 
python-dotenv